отрефаторила utils.js, разделила на 3 файла
