+++ src/mock/task.js
@@ -1,5 +1,5 @@
 import dayjs from 'dayjs';
-import { getRandomInteger } from '../utils';
+import { getRandomInteger } from '../utils/common';
 import { getOffer, generatePlace, generateDate } from './mocks';
 
 export const generateTask = () => {
