+++ src/view/offers-selector.js
@@ -1,4 +1,4 @@
-import { createElement } from '../utils';
+import { createElement } from '../utils/render';
 
 const createOffersSelectorTemplate = (task) => {
   const { offers } = task;
